+++ CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4] - 2020-11-22
+
+### Added
+- Added storage permission to manifest
+- Added popup menu
+- Added stored options in browser
+- Added Semantic UI for popup styles
+
+### Changed
+- Moved content to folder
+- Moved popup to folder
+- Moved polls filtering to their own function
+- Cleaned some code
+
 ## [0.3] - 2020-11-21
 
 ### Added